Move Node functionality from Entity to Node
This cleanup seems to have been forgotten when the Entity was
introduced to replace the former Node that was parent to Edge.
